add FringeZernike coefficient file for the jas36xm2 model
The orignal `FringeZernike_2D-s1.csv` file is renamed to
`FringeZernike_2D-s1_20211130.csv` based on the date when the file was
added to this repository.

A new Fringe Zernike wavefront error file `FringeZernike_2D-s1_jas36xm2.csv`
is added. This wavefront error is calculated based on the jas36xm2 model.

A symbolic link is created to select which wavefront error file is used
in the simulation. In this commit, the symbolic link is pointed to the
wavefront error of the jas36xm2 model.